Button labels still falling behind
Got the same functionality with for instead of foreach--the buttons work but have the wrong labels. Seem to reliably be one button label behind the times (though first screen with options just has the two choices reversed). Having never mastered lists I can't get that label to be for the i+1 without being "outside the array" and I don't know how else to approach this.